+++ src/ncmpcpp.cpp
@@ -103,7 +103,7 @@ int main(int argc, char **argv)
 	// always execute these commands, even if ncmpcpp use exit function
 	atexit(do_at_exit);
 	
-	// redirect std::cerr output to ~/.ncmpcpp/error.log file
+	// redirect std::cerr output to the error.log file
 	errorlog.open((Config.ncmpcpp_directory + "error.log").c_str(), std::ios::app);
 	cerr_buffer = std::cerr.rdbuf();
 	std::cerr.rdbuf(errorlog.rdbuf());