+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * Deprecate `visualizer_sync_interval` configuration option.
 * Deprecate `noop` value of `browser_sort_mode` in favor of `none`.
 * Add `type` value of `browser_sort_mode` (set by default).
+* Prefer `$XDG_CONFIG_HOME/ncmpcpp` as a folder with configuration.
 
 # ncmpcpp-0.8.2 (2018-04-11)
 * Help screen: fixed display of EoF keycode